pciutils: lspci: fix two small bug reports
The sanitizer patch found two small issues as follows:
lspci.c:198:23: runtime error: left shift of 255 by 24 places cannot be
represented in type 'int'
ls-ecaps.c:561:2: runtime error: left shift of 1 by 31 places cannot be
represented in type 'int'

Fix both these issues by converting the data being shifted to an
unsigned long or u32 before shifting.

+++ lspci.c
@@ -196,9 +196,9 @@ get_conf_long(struct device *d, unsigned int pos)
 {
   check_conf_range(d, pos, 4);
   return d->config[pos] |
-    (d->config[pos+1] << 8) |
-    (d->config[pos+2] << 16) |
-    (d->config[pos+3] << 24);
+    ((u32)(d->config[pos+1]) << 8) |
+    ((u32)(d->config[pos+2]) << 16) |
+    ((u32)(d->config[pos+3]) << 24);
 }
 
 /*** Sorting ***/
